requirements.txt
gradio
torch
torchvision
numpy
pillow
requests

